src/test/java/com/leetcode/algorithm/medium/verticalordertraversal/SolutionTest.java
package com.leetcode.algorithm.medium.verticalordertraversal;

import com.leetcode.algorithm.common.TreeNode;
import org.junit.jupiter.api.Test;

import java.util.Arrays;

import static org.junit.jupiter.api.Assertions.assertEquals;

public class SolutionTest {
  @Test
  public void testCase1() {
    Solution solution = new Solution();
    TreeNode root = new TreeNode(3);
    root.left = new TreeNode(9);
    root.right = new TreeNode(20);
    root.right.left = new TreeNode(15);
    root.right.right = new TreeNode(7);

    assertEquals(
        Arrays.asList(Arrays.asList(9), Arrays.asList(3, 15), Arrays.asList(20), Arrays.asList(7)),
        solution.verticalTraversal(root));
  }

  @Test
  public void testCase2() {
    Solution solution = new Solution();
    TreeNode root = new TreeNode(1);
    root.left = new TreeNode(2);
    root.left.left = new TreeNode(4);
    root.left.right = new TreeNode(5);
    root.right = new TreeNode(3);
    root.right.left = new TreeNode(6);
    root.right.right = new TreeNode(7);

    assertEquals(
        Arrays.asList(
            Arrays.asList(4),
            Arrays.asList(2),
            Arrays.asList(1, 5, 6),
            Arrays.asList(3),
            Arrays.asList(7)),
        solution.verticalTraversal(root));
  }
}
